+++ clientcontroller/api/interface.go
@@ -106,7 +106,7 @@ type FinalityOperator interface {
 
 type SubmitBatchFinalitySigsRequest struct {
 	FpPk        *btcec.PublicKey
-	Blocks      []*types.BlockInfo
+	Blocks      []types.BlockDescription
 	PubRandList []*btcec.FieldVal
 	ProofList   [][]byte
 	Sigs        []*btcec.ModNScalar
@@ -138,7 +138,7 @@ type QueryFinalityProviderHasPowerRequest struct {
 
 func NewSubmitBatchFinalitySigsRequest(
 	fpPk *btcec.PublicKey,
-	blocks []*types.BlockInfo,
+	blocks []types.BlockDescription,
 	pubRandList []*btcec.FieldVal,
 	proofList [][]byte,
 	sigs []*btcec.ModNScalar,